# Fixed #35518 -- Used simpler string operations for converter-less routes