--- a/tests/urlpatterns/test_resolvers.py
+++ b/tests/urlpatterns/test_resolvers.py
@@ -13,6 +13,12 @@ class RoutePatternTests(SimpleTestCase):
     def test_str(self):
         self.assertEqual(str(RoutePattern(_("translated/"))), "translated/")
 
+    def test_has_converters(self):
+        self.assertEqual(len(RoutePattern("translated/").converters), 0)
+        self.assertEqual(len(RoutePattern(_("translated/")).converters), 0)
+        self.assertEqual(len(RoutePattern("translated/<int:foo>").converters), 1)
+        self.assertEqual(len(RoutePattern(_("translated/<int:foo>")).converters), 1)
+
 
 class ResolverCacheTests(SimpleTestCase):
     @override_settings(ROOT_URLCONF="urlpatterns.path_urls")