@RealOrangeOne 2025-05-02
I've added some more benchmarks to [the ticket](https://code.djangoproject.com/ticket/35518#comment:8) which make the gain clearer.